requirements.txt
flask==3.0.3
python-dotenv==1.0.1
qrcode==7.4.2
supabase==2.6.0